+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * **build**: update node dependencies
 * **misc**: Add provisioning files
 * **config**: Update appconfig to be in line with plugin best practices
+* **queries**: Add default queries for the datasources
 
 ## 6.0.1 (2026-05-20)